Moved classed to a onesided package.
